// src/app/model/commessaCliente.ts
export class CommessaCliente{
    _id: any;
    id_cliente: any;
    nome_cliente: string;
    id_commessa_fnc: any;
    nome_commessa_fnc: string;
    codice_commessa: string;
    nome_commessa: string;
    budget_euro: Number;
    budget_gg: Number;
    codice_offerta: string;
    codice_ordine: string;
    data_inizio_validita: Date;
    data_fine_validita: Date;
}